// src/api/B2C/travelers/travelers.interfaces.ts
export interface ITravellersBody {
  user_id?: number;
  PassengerType: string;
  Gender: string;
  PassengerTitle: string;
  PassengerFirstName: string;
  PassengerLastName: string;
  DateOfBirth: string;
  PassengerNationality: string;
  NationalID?: string;
  PassportNumber?: string;
  ExpiryDate?: string;
  Country: string;
}